Update coll framework count/disp arrays for bigcount
This updates the coll framework functions using count/displacement
arrays to support bigcount. Instead of directly using pointers to
bigcount/non-bigcount type arrays, this adds special descriptor types,
ompi_count_array_t and ompi_disp_array_t, that internally hold a union of
either type. Collective components can now access count and displacement
values through inline get functions on the descriptors, allowing use of
both bigcount/non-bigcount arrays, depending on how the descriptors were
initialized.

+++ ompi/mca/coll/portals4/coll_portals4.h
@@ -222,26 +222,26 @@ int ompi_coll_portals4_iallreduce_intra(const void* sendbuf, void* recvbuf, int 
 int
 ompi_coll_portals4_iallreduce_intra_fini(struct ompi_coll_portals4_request_t *request);
 
-int ompi_coll_portals4_gather_intra(const void *sbuf, int scount, struct ompi_datatype_t *sdtype,
-                                    void *rbuf, int rcount, struct ompi_datatype_t *rdtype,
+int ompi_coll_portals4_gather_intra(const void *sbuf, size_t scount, struct ompi_datatype_t *sdtype,
+                                    void *rbuf, size_t rcount, struct ompi_datatype_t *rdtype,
                                     int root,
                                     struct ompi_communicator_t *comm,
                                     mca_coll_base_module_t *module);
-int ompi_coll_portals4_igather_intra(const void *sbuf, int scount, struct ompi_datatype_t *sdtype,
-                                     void *rbuf, int rcount, struct ompi_datatype_t *rdtype,
+int ompi_coll_portals4_igather_intra(const void *sbuf, size_t scount, struct ompi_datatype_t *sdtype,
+                                     void *rbuf, size_t rcount, struct ompi_datatype_t *rdtype,
                                      int root,
                                      struct ompi_communicator_t *comm,
                                      ompi_request_t **request,
                                      mca_coll_base_module_t *module);
 int ompi_coll_portals4_igather_intra_fini(struct ompi_coll_portals4_request_t *request);
 
-int ompi_coll_portals4_scatter_intra(const void *sbuf, int scount, struct ompi_datatype_t *sdtype,
-                                     void *rbuf, int rcount, struct ompi_datatype_t *rdtype,
+int ompi_coll_portals4_scatter_intra(const void *sbuf, size_t scount, struct ompi_datatype_t *sdtype,
+                                     void *rbuf, size_t rcount, struct ompi_datatype_t *rdtype,
                                      int root,
                                      struct ompi_communicator_t *comm,
                                      mca_coll_base_module_t *module);
-int ompi_coll_portals4_iscatter_intra(const void *sbuf, int scount, struct ompi_datatype_t *sdtype,
-                                      void *rbuf, int rcount, struct ompi_datatype_t *rdtype,
+int ompi_coll_portals4_iscatter_intra(const void *sbuf, size_t scount, struct ompi_datatype_t *sdtype,
+                                      void *rbuf, size_t rcount, struct ompi_datatype_t *rdtype,
                                       int root,
                                       struct ompi_communicator_t *comm,
                                       ompi_request_t **request,
